+++ examples/pipeline/_02_full_pipeline.py
@@ -7,14 +7,17 @@ Note, that we provide pre-built pipelines for common use-cases in the package.
 Checkout the examples for those, if you want to understand how to use them.
 
 """
-# TODO: This example is WIP and will be expanded once the respective algorithms are implemented.
 
+import pandas as pd
+
+# TODO: This example is WIP and will be expanded once the respective algorithms are implemented.
 # %%
 # Load example data
 # -----------------
 # We load example data from the lab dataset, and we will use a single long-trail from a "MS" participant for this
 # example.
 from mobgap.data import LabExampleDataset
+from mobgap.utils.interpolation import naive_sec_paras_to_regions
 
 lab_example_data = LabExampleDataset(reference_system="INDIP")
 long_trial = lab_example_data.get_subset(
@@ -90,7 +93,7 @@ cad.calculate(
     sampling_rate_hz=sampling_rate_hz,
 )
 
-cad_per_sec = cad.cad_per_sec_
+cad_per_sec = cad.cadence_per_sec_
 cad_per_sec
 
 # %%
@@ -140,7 +143,7 @@ for (_, gs_data), r in gs_iterator.iterate(imu_data, gait_sequences):
         cad.calculate(
             refined_gs_data, refined_ic_list, sampling_rate_hz=sampling_rate_hz
         )
-        rr.cad_per_sec = cad.cad_per_sec_
+        rr.cadence_per_sec = cad.cadence_per_sec_
 
 # %%
 # Now we can access all accumulated and offset-corrected results from the iterator.
@@ -149,20 +152,140 @@ results = gs_iterator.results_
 results.ic_list
 
 # %%
-results.cad_per_sec
+results.cadence_per_sec
 
 # %%
 # Using the combined results, we want to define walking bouts.
 # As walking bouts in the context of Mobilise-D are defined based on strides, we need to turn the ICs into strides and
 # the per-second values into per-stride values by using interpolation.
-# TODO: This is not implemented yet.
-# stride_list = stride_from_ic_list(results.ic_list)
-# stride_list_with_paras = interpolate_per_stride(
-#     stride_list, cad_per_sec=results.cad_per_sec, stride_length=results.stride_length
-# )
+from mobgap.lrc import strides_list_from_ic_lr_list
+
+stride_list = results.ic_list.groupby("gs_id", group_keys=False).apply(
+    strides_list_from_ic_lr_list
+)
+stride_list
+# %%
+# This initial stride list is completely unfiltered, and might contain very long strides, in areas where initial
+# contacts were not detected, or the participant was not walking for a short moment.
+# The stride list will be filtered later as part of the WB assembly.
 #
+# For now, we are using linear interpolation to map the per-second cadence values to per-stride values and derive
+# approximated stride parameters.
+# We also calculate the stride duration here.
 #
-# from mobgap.wba import StrideSelection, WbAssembly
+# Note, that we remove the ``r_gs_id`` index, as we don't need it anymore and each normal ``gs`` is mapped to a single
+# refined ``gs`` anyway.
+# In case we would have multiple refined ``gs`` per normal ``gs``, we might need to keep the ``r_gs_id`` index around.
+from mobgap.utils.array_handling import create_multi_groupby
+
+combined_results = pd.concat(
+    [
+        results.cadence_per_sec,
+        results.stride_length_per_sec,
+        results.walking_speed_per_sec,
+    ],
+    axis=1,
+).reset_index("r_gs_id", drop=True)
+
+stride_list_with_approx_paras = (
+    create_multi_groupby(
+        stride_list,
+        combined_results,
+        "gs_id",
+        group_keys=False,
+    )
+    .apply(naive_sec_paras_to_regions, sampling_rate_hz=sampling_rate_hz)
+    .assign(
+        stride_duration_s=lambda df_: (df_.end - df_.start) / sampling_rate_hz
+    )
+)
+
+stride_list_with_approx_paras
+# %%
+# Now the final strides are regrouped into walking bouts.
+# For this we ignore which gait sequence the strides belong to, hence we remove the ``gs_id`` from the index, but keep
+# it around as column for debugging.
+from mobgap.wba import StrideSelection, WbAssembly
+
+stride_list_with_approx_paras = stride_list_with_approx_paras.reset_index(
+    "gs_id"
+).rename(columns={"gs_id": "original_gs_id"})
+
+# %%
+# Then we apply the stride selection (note that we have additional rules in case the stride length is available) and
+# then group the remaining strides into walking bouts.
+ss_rules = (
+    StrideSelection.PredefinedParameters.mobilised
+    if "stride_length_m" in stride_list_with_approx_paras.columns
+    else StrideSelection.PredefinedParameters.mobilised_no_stride_length
+)
+ss = StrideSelection(**ss_rules).filter(
+    stride_list_with_approx_paras, sampling_rate_hz=sampling_rate_hz
+)
+wba = WbAssembly().assemble(
+    ss.filtered_stride_list_, sampling_rate_hz=sampling_rate_hz
+)
+
+final_strides = wba.annotated_stride_list_
+final_strides
+
+# %%
+# We also have meta information about the WBs available.
+per_wb_params = wba.wb_meta_parameters_
+per_wb_params
+
+# %%
+# We extend them further with the per-stride parameters.
+params_to_aggregate = [
+    "stride_duration_s",
+    "cadence_spm",
+    "stride_length_m",
+    "walking_speed_mps",
+]
+per_wb_params = pd.concat(
+    [
+        per_wb_params,
+        final_strides.reindex(columns=params_to_aggregate)
+        .groupby(["wb_id"])
+        # TODO: Decide if we should use mean or trim_mean here!
+        .mean(),
+    ],
+    axis=1,
+)
+
+per_wb_params
+
+# %%
+# For each WB we can then apply thresholds to check if the calculated parameters are within the expected range.
+from mobgap.aggregation import apply_thresholds, get_mobilised_dmo_thresholds
+
+thresholds = get_mobilised_dmo_thresholds()
+
+per_wb_params_mask = apply_thresholds(
+    per_wb_params,
+    thresholds,
+    cohort=long_trial.participant_metadata["cohort"],
+    height_m=long_trial.participant_metadata["height_m"],
+    measurement_condition=long_trial.recording_metadata[
+        "measurement_condition"
+    ],
+)
+per_wb_params_mask
+
+# %%
+# We can see that we either get NaN (for parameters that are not checked) or True/False values for each parameter.
 #
-# ss = StrideSelection().filter()
-# wba = WbAssembly().assemble()
+# This output together with the per-WB parameters would then normally be used in some aggregation step to calculate
+# single values per participant, day, or other grouping criteria.
+# Depending on the use-case, this aggregation can be performed withing the "per-recording" pipeline or as a separate
+# step after processing all recordings.
+#
+# Here, we perform it per recording and calculate a single values from all the WBs.
+from mobgap.aggregation import MobilisedAggregator
+
+agg_results = (
+    MobilisedAggregator(groupby=None)
+    .aggregate(per_wb_params, wb_dmos_mask=per_wb_params_mask)
+    .aggregated_data_
+)
+agg_results